feed level

+++ FeederInfo.h
@@ -2,18 +2,21 @@
 #ifndef FEEDER_INFO_H
 #define FEEDER_INFO_H
 
-#include <vector>  
+#include <vector>
 #include <Arduino.h>
 
-struct Feeder {
+struct Feeder
+{
   String name;
   String status;
   String ipAddress;
-  float levelLeft;
-  float levelRight;
+  // food level min:110mm max:90mm
+  uint32_t levelLeft;
+  uint32_t levelRight;
 };
 
-struct Schedule {
+struct Schedule
+{
   String UUID;
   int side;
   int amount;
